requirements.txt
asgiref==3.2.10
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-heroku==0.3.1
django-model-utils==4.0.0
djangorestframework==3.12.0
psycopg2==2.8.6
pytubeX==0.1.2
pytz==2020.1
sqlparse==0.3.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
